refactor(kitchenpos/common): move common domain model test codes
