+++ src/test/java/kitchenpos/common/tobe/domain/model/PriceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.tobe;
+package kitchenpos.common.tobe.domain.model;
 
 import kitchenpos.commons.tobe.domain.model.Price;
 import org.junit.jupiter.api.DisplayName;
