+++ src/test/java/kitchenpos/common/tobe/domain/model/QuantityTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.tobe;
+package kitchenpos.common.tobe.domain.model;
 
 import kitchenpos.commons.tobe.domain.model.Quantity;
 import org.junit.jupiter.api.DisplayName;
